+++ extensions/pi-flows/parse.ts
@@ -1,6 +1,5 @@
 export {
 	extractLastJsonBlock,
-	isPassWord,
 	parseLoopStatus,
 	parseRoute,
 	parseScore,